pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "neurosearch"
version = "0.1.0"
description = "Hybrid Generative Retrieval & RAG for e-commerce search using the Amazon ESCI dataset"
readme = "README.md"
authors = [{name = "Your Name", email = "you@example.com"}]
license = {text = "MIT"}
requires-python = ">=3.9"
dependencies = [
    "sentence-transformers",
    "faiss-cpu",
    "torch",
    "transformers",
    "pandas",
    "pyarrow",
    "scikit-learn",
    "datasets",
    "tqdm",
    "kagglehub",
    "rank_bm25",
    "matplotlib",
    "seaborn"
]

[tool.setuptools.packages.find]
where = ["src"]
